[FIN-349] feat: 카테고리 목록 API

+++ src/main/java/kr/co/finote/backend/src/article/service/ArticleService.java
@@ -306,7 +306,7 @@ public class ArticleService {
         return articleRepository.existsByCategoryAndIsDeleted(category, false);
     }
 
-    public int countByCategoryAndUser(Category category, User loginUser) {
+    public Long countByCategoryAndUser(Category category, User loginUser) {
         return articleRepository.countAllByCategoryAndUserAndIsDeleted(category, loginUser, false);
     }
 }
